Created "Backup - HLMV Rotations Page - Files" folder.
This is a backup / archive of all the files from the HLMV Rotations Page (except the ones created on this repo).

In case any of these files get deleted by their original uploader(s), or if the service that these files were stored in, is shutdown.

It also includes some that are no longer on the page.