Some fixes
The Magical Mount and Barefoot Brawler files are now in .zips, due to requiring the user to change the color settings for paint variants. Making it a .vpk only adds an extra step that requires the user to extract via a third-party tool.

Also fixed an error with the Down Under Duster Files materials.